Basic Mic Behaviour
Can be picked up, shows up on HUD, grant points to the team carrying it, drop on death. The code should allow for any number of team in a game but is currently hard coded for two. Also, there is a small bug that sometimes occurs, causing the Client and Server score to differ by 1.