Code Clean-up
Cleaned the HUD, Beat Widgets are now recycled going from 618 total widgets created to 32. Microphone is no longer destroyed by the player.  Gold Areas can't stay on the same spot when moving. Every blueprints was checked and now all are up to the Codex Unrealis' latest standards.

New bugs : Discovered that Area Movement is not replicated, the Mic behaves weirdly when dropped which is most likely due to the replication bug of Area movement.