Death Logic Working
Spectate a Player on Death, switch Player if that Player dies, if no Player are available spectate corpse, if a Player spawns while spectating corpse, switch to spectate that Player. Multiple Spectators are supported on the same Player.
Need to map out the logic of the whole project, things are starting to get messy and confusing.
Health bar is not replicated properly.